+++ NEWS.md
@@ -1,3 +1,5 @@
+# microViz (development version)
+
 # microViz 0.12.5
 
 - Fix: tax_fix no longer allows min_length of 0, as empty strings are (and were) always replaced